Update visual-c++\B_spline - Porting to VC++ 2017 - rev4
- Change vcxproj name from LAGRANGE to B_spline
- Added a new icon to the project
- Add credits and version history
- Translate GUI from spanish to english
- Warnings due to lack of static_cast are corrected

+++ visual-c++/B_spline/LAGRANGEDlg.h
@@ -20,7 +20,7 @@ public:
 	double t[6];
 	double angulo[2];
 	double p[3][3];
-	double pc[2][2];
+	int pc[2][2];
 	double c[2];
 	double puntoaux[3];
 	double punto[3][3];
